requirements.txt
requests>=2.31.0
aiohttp>=3.8.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.7.0
mistralai>=0.0.8
together>=0.2.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.25.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
